Separate userspace headers and eBPF programs
