+++ pkg/ebpf/c/oom-kill-kern.c
@@ -2,7 +2,7 @@
 #include <linux/oom.h>
 
 #include "pkg/ebpf/c/bpf-common.h"
-#include "pkg/ebpf/c/oom-kill-kern-user.h"
+#include "pkg/ebpf/oom-kill-kern-user.h"
 
 
 /*
