+++ pkg/ebpf/c/tcp-queue-length-kern.c
@@ -6,7 +6,7 @@
 #include <linux/tcp.h>
 
 #include "pkg/ebpf/c/bpf-common.h"
-#include "pkg/ebpf/c/tcp-queue-length-kern-user.h"
+#include "pkg/ebpf/tcp-queue-length-kern-user.h"
 
 /*
  * The `queue` map is used to share with the userland program system-probe
